Supported ExternalTrafficPolicy for CreateK8sService.

+++ aliyun-java-sdk-edas/src/main/java/com/aliyuncs/edas/model/v20170801/AddMockRuleRequest.java
@@ -45,7 +45,7 @@ public class AddMockRuleRequest extends RoaAcsRequest<AddMockRuleResponse> {
 
 	private String consumerAppsJson;
 
-	private Long mockType;
+	private Integer mockType;
 
 	private String region;
 	public AddMockRuleRequest() {
@@ -168,11 +168,11 @@ public class AddMockRuleRequest extends RoaAcsRequest<AddMockRuleResponse> {
 		}
 	}
 
-	public Long getMockType() {
+	public Integer getMockType() {
 		return this.mockType;
 	}
 
-	public void setMockType(Long mockType) {
+	public void setMockType(Integer mockType) {
 		this.mockType = mockType;
 		if(mockType != null){
 			putQueryParameter("MockType", mockType.toString());
