+++ aliyun-java-sdk-edas/src/main/java/com/aliyuncs/edas/model/v20170801/GetAccountMockRuleRequest.java
@@ -35,7 +35,7 @@ public class GetAccountMockRuleRequest extends RoaAcsRequest<GetAccountMockRuleR
 
 	private String namespace;
 
-	private Long mockType;
+	private Integer mockType;
 
 	private String region;
 
@@ -105,11 +105,11 @@ public class GetAccountMockRuleRequest extends RoaAcsRequest<GetAccountMockRuleR
 		}
 	}
 
-	public Long getMockType() {
+	public Integer getMockType() {
 		return this.mockType;
 	}
 
-	public void setMockType(Long mockType) {
+	public void setMockType(Integer mockType) {
 		this.mockType = mockType;
 		if(mockType != null){
 			putQueryParameter("MockType", mockType.toString());
